move enums and schema to top level of package
