+++ mndot_bid_api/routers/bidders.py
@@ -1,6 +1,6 @@
 import fastapi
-from mndot_bid_api import db, operations
-from mndot_bid_api.operations import schema
+
+from mndot_bid_api import db, operations, schema
 
 router = fastapi.APIRouter(prefix="/bidder", tags=["bidder"])
