+++ mndot_bid_api/routers/contracts.py
@@ -1,8 +1,8 @@
 from datetime import date
 
 import fastapi
-from mndot_bid_api import db, operations
-from mndot_bid_api.operations import schema
+
+from mndot_bid_api import db, operations, schema
 
 router = fastapi.APIRouter(prefix="/contract", tags=["contract"])
