+++ mndot_bid_api/routers/items.py
@@ -1,7 +1,6 @@
 import fastapi
 
-from mndot_bid_api import db, operations
-from mndot_bid_api.operations import schema
+from mndot_bid_api import db, operations, schema
 
 router = fastapi.APIRouter(prefix="/item", tags=["item"])
